Re-organized the QNX support.
As most of the differences are related to the change in network stack
half-way through the life of QNX Neutrino 7.1, the definitions are now
split into io-pkt (the older stack) and io-sock (the newer stack).

I didn't add any architecture definitions for x86 (as would be required
for i686-pc-nto-qnx700) because I don't have access to QNX Neutrino 7.0
to verify what should go here. But there's a clear place to add it
later.

I do have QNX Neutrino 7.1 and QNX OS 8.0 and I have verified each of
the types in the io-pkt and io-sock modules against the relevant C
header files. I have not verified the remaining types or functions - I
assume they are correct.
