+++ src/unix/nto/arch/aarch64.rs
@@ -1,3 +1,12 @@
+//! Definitions specific to QNX on AArch64
+//!
+//! This module applies to:
+//!
+//! * `aarch64-unknown-nto-qnx700`
+//! * `aarch64-unknown-nto-qnx710`
+//! * `aarch64-unknown-nto-qnx710_iosock`
+//! * `aarch64-unknown-nto-qnx800`
+
 use crate::prelude::*;
 
 pub type wchar_t = u32;
