+++ src/unix/nto/arch/i686.rs
@@ -1,0 +1,10 @@
+//! Definitions specific to QNX on x86
+//!
+//! This module applies to:
+//!
+//! * `i686-pc-nto-qnx700`
+
+use crate::prelude::*;
+
+pub type wchar_t = u32;
+pub type time_t = i64;
