+++ src/unix/nto/arch/mod.rs
@@ -1,0 +1,13 @@
+//! Architecture-specific definitions for QNX
+
+cfg_if! {
+    if #[cfg(target_arch = "x86_64")] {
+        mod x86_64;
+        pub use self::x86_64::*;
+    } else if #[cfg(target_arch = "aarch64")] {
+        mod aarch64;
+        pub use self::aarch64::*;
+    } else {
+        panic!("Unsupported arch");
+    }
+}
