+++ src/unix/nto/arch/x86_64.rs
@@ -1,3 +1,11 @@
+//! Definitions specific to QNX on x86-64
+//!
+//! This module applies to:
+//!
+//! * `x86_64-pc-nto-qnx710`
+//! * `x86_64-pc-nto-qnx710_iosock`
+//! * `x86_64-pc-nto-qnx800`
+
 use crate::prelude::*;
 
 pub type wchar_t = u32;
